requirements.txt
flask
flask-socketio
eventlet
gevent
mysql-connector-python
imap-tools
selenium
webdrivermanager